Create widgetCollectionView
